게시물 API

+++ src/main/java/org/example/board/service/PostService.java
@@ -1,6 +1,8 @@
 package org.example.board.service;
 
 import org.example.board.exception.post.PostNotFoundException;
+import org.example.board.exception.user.UserNotAllowedException;
+import org.example.board.model.entity.UserEntity;
 import org.example.board.model.post.Post;
 import org.example.board.model.post.PostPatchRequestBody;
 import org.example.board.model.post.PostPostRequestBody;
@@ -32,19 +34,22 @@ public class PostService {
     }
 
 
-    public Post createPost(PostPostRequestBody postPostRequestBody) {
-        PostEntity postEntity = new PostEntity();
-        postEntity.setBody(postPostRequestBody.body());
-        var savedPostEntity = postEntityRepository.save(postEntity);
-
-        return Post.from(savedPostEntity);
+    public Post createPost(PostPostRequestBody postPostRequestBody, UserEntity currentUser) {
+        var postEntity = postEntityRepository.save(
+                PostEntity.of(postPostRequestBody.body(), currentUser)
+        );
+        return Post.from(postEntity);
     }
 
-    public Post updatePost(Long postId, PostPatchRequestBody postPatchRequestBody) {
+    public Post updatePost(Long postId, PostPatchRequestBody postPatchRequestBody, UserEntity currentUser) {
         var postEntity = postEntityRepository
                 .findById(postId)
                 .orElseThrow(
                         () -> new PostNotFoundException(postId));
+
+        if(!postEntity.getUser().equals(currentUser)){
+            throw new UserNotAllowedException();
+        }
 
         postEntity.setBody(postPatchRequestBody.body());
         var updatedPostEntity = postEntityRepository.save(postEntity);
@@ -53,11 +58,15 @@ public class PostService {
     }
 
 
-    public void deletePost(Long postId) {
+    public void deletePost(Long postId, UserEntity currentUser) {
         var postEntity = postEntityRepository
                 .findById(postId)
                 .orElseThrow(
                         () -> new PostNotFoundException(postId));
+
+        if(!postEntity.getUser().equals(currentUser)){
+            throw new UserNotAllowedException();
+        }
 
         postEntityRepository.delete(postEntity);
     }
